Extracted progr.h from edit.h. Subsequently extracted several headers from progr.h. Finally solved some of the old compiler warning (with lowest warning level of gcc).
