+++ Makefile.in
@@ -86,7 +86,7 @@ CFILES=		we_main.c we_block.c we_unix.c we_e_aus.c \
 		WeXterm.c WeSyntax.c WeExpArr.c WeLinux.c we_gpm.c
 
 HFILES=		attrb.h edit.h keys.h \
-		model.h progr.h unixkeys.h unixmakr.h \
+		model.h unixkeys.h unixmakr.h \
 		Xwpe.h WeString.h WeXterm.h WeProg.h WeExpArr.h
 
 xwpe:	$(OFILES) $(X_OFILES) $(T_OFILES)
