+++ we_fl_unix.h
@@ -11,11 +11,8 @@
 
 #ifdef UNIX
 #include "we_fl_unix_private.h"
-/** Commented out #endif and placed after prototypes,
- * because the functions for the following prototypes in
- * the we_fl_unix.c source are all included under #ifdef UNIX.
- */
-//#endif  
+
+int e_data_eingabe(ECNT *cn);
 
 char *WpeGetCurrentDir(ECNT *cn);
 struct dirfile *WpeCreateWorkingDirTree(int sw, ECNT *cn);
@@ -24,6 +21,13 @@ char *WpeAssemblePath(char *pth, struct dirfile *cd, struct dirfile *dd, int n,
 struct dirfile *WpeGraphicalFileList(struct dirfile *df, int sw, ECNT *cn);
 struct dirfile *WpeGraphicalDirTree(struct dirfile *cd, struct dirfile *dd,
                                     ECNT *cn);
+/*   we_fl_unix.c  */
+
+int e_funct(FENSTER *f);
+int e_funct_in(FENSTER *f);
+int e_data_first(int sw, ECNT *cn, char *nstr);
+int e_data_schirm(FENSTER *f);
+
 #endif
 
 #endif
