+++ we_xterm.c
@@ -7,6 +7,7 @@
 #ifndef NO_XWINDOWS
 
 #include "edit.h"
+#include "we_unix.h"
 #include "we_xterm.h"
 
 /* partial conversion in place */
@@ -16,7 +17,6 @@
 #define WpeDllInit WpeXtermInit
 #endif
 
-int e_X_sw_color(void);
 int fk_show_cursor(void);
 int e_ini_size(void);
 int e_x_getch(void);
@@ -96,7 +96,7 @@ int WpeDllInit(int *argc, char **argv)
  e_u_ini_size = e_ini_size;
  e_u_setlastpic = e_setlastpic;
  WpeMouseChangeShape = (void (*)(WpeMouseShape))WpeNullFunction;
- WpeMouseRestoreShape = (void (*)(WpeMouseShape))WpeNullFunction;
+ WpeMouseRestoreShape = (void (*)(void))WpeNullFunction;
 /* WpeMouseChangeShape = WpeXMouseChangeShape;
  WpeMouseRestoreShape = WpeXMouseRestoreShape;*/
  WpeDisplayEnd = WpeNullFunction;
